Refactoring Gulpfile using gulp-di
