+++ gulp/modules/change-event.js
@@ -1,0 +1,13 @@
+module.exports = function (log, config) {
+  /**
+   * When files change, log it
+   * @param  {Object} event - event that fired
+   */
+  function changeEvent(event) {
+      var srcPattern = new RegExp('/.*(?=/' + config.source + ')/');
+      log('File ' + event.path.replace(srcPattern, '') + ' ' + event.type);
+  }
+
+  return changeEvent;
+
+};
