+++ gulp/modules/error-logger.js
@@ -1,0 +1,13 @@
+module.exports = function (log) {
+	/**
+	 * Log an error message and emit the end of a task
+	 */
+	function errorLogger(error) {
+			log('*** Start of Error ***');
+			log(error);
+			log('*** End of Error ***');
+			this.emit('end');
+	}
+
+	return errorLogger;
+};
