+++ gulp/modules/run-node-inspector.js
@@ -1,0 +1,12 @@
+module.exports = function (log) {
+
+  function runNodeInspector() {
+      log('Running node-inspector.');
+      log('Browse to http://localhost:8080/debug?port=5858');
+      var exec = require('child_process').exec;
+      exec('node-inspector');
+  }
+
+  return runNodeInspector;
+
+};
