+++ gulp/tasks/clean-images.js
@@ -1,0 +1,10 @@
+module.exports = function (gulp, config, clean) {
+
+    gulp.task('clean-images', function(done) {
+        /**
+         * Remove all images from the build folder
+         * @param  {Function} done - callback when complete
+         */
+        clean(config.build + 'images/**/*.*', done);
+    });
+};
