+++ gulp/tasks/default.js
@@ -1,0 +1,3 @@
+module.exports = function (gulp) {
+    gulp.task('default', ['help']);
+};
