+++ gulp/tasks/less-watcher.js
@@ -1,0 +1,6 @@
+module.exports = function (gulp, config) {
+
+    gulp.task('less-watcher', function() {
+        gulp.watch([config.less], ['styles']);
+    });
+};
